requirements.txt
python-rtmidi==1.4.6
keyboard==0.13.5
